feat: update ID types from uint to uint64 across models, services, and repositories for improved scalability

+++ internal/audit/admin.go
@@ -6,8 +6,8 @@ import (
 
 type AdminAuditParams struct {
 	Action           string
-	TargetUserID     uint
-	PerformedByID    uint
+	TargetUserID     uint64
+	PerformedByID    uint64
 	PerformedByEmail string
 	Success          bool
 	ErrorMsg         string
