+++ internal/dto/theater/theater_response.go
@@ -1,6 +1,6 @@
 package theater
 
 type TheaterResponse struct {
-	ID   uint   `json:"id"`
+	ID   uint64 `json:"id"`
 	Name string `json:"name"`
 }
