+++ internal/repositories/movie_repo.go
@@ -17,7 +17,7 @@ func (r *MovieRepository) CreateMovie(movie *models.Movie) error {
 	return r.db.Create(movie).Error
 }
 
-func (r *MovieRepository) FindById(movieID uint) (*models.Movie, error) {
+func (r *MovieRepository) FindById(movieID uint64) (*models.Movie, error) {
 	var movie models.Movie
 	if err := r.db.First(&movie, movieID).Error; err != nil {
 		return nil, err
